test: auto-select Vulkan in ocr-ggml integration harness when a Vulkan backend lib is present
Factor findVulkanBackendLib() and PREBUILDS_DIR out of
backend-device.test.js into utils.js and export them. backend-device.test.js
now imports both (behaviour unchanged).

Extend getBackendDevice() precedence:
  1. explicit OCR_GGML_BACKEND wins (vulkan only when exactly `vulkan`,
     else cpu) — preserves manual override.
  2. else, on desktop, auto-select vulkan when a ggml-vulkan backend lib
     is shipped in prebuilds/. On desktop CI the merged prebuilds contain
     the lib, so the full EasyOCR/DocTR suites attempt Vulkan inside the
     existing run-integration-tests job; only the GPU runner actually
     executes on Vulkan, others report an explicit CPU fallback. Local dev
     without merged prebuilds stays on CPU.
  3. else cpu.

Update README to document the auto-detect behaviour and that
OCR_GGML_BACKEND remains a manual override.
